compse_err moved to _utils
