+++ nilearn/_utils/__init__.py
@@ -6,3 +6,5 @@ from .numpy_conversions import as_ndarray
 
 from .cache_mixin import CacheMixin
 
+from .logger import _compose_err_msg
+
