+++ nilearn/input_data/nifti_maps_masker.py
@@ -13,32 +13,10 @@ from .. import _utils
 from .._utils import logger
 from .._utils import CacheMixin
 from .._utils.niimg_conversions import _check_same_fov
+from .._utils import _compose_err_msg
 from .. import signal
 from .. import region
 from .. import image
-
-
-def _compose_err_msg(msg, **kwargs):
-    """Append key-value pairs to msg, for display.
-
-    Parameters
-    ==========
-    msg: string
-        arbitrary message
-    kwargs: dict
-        arbitrary dictionary
-
-    Returns
-    =======
-    updated_msg: string
-        msg, with "key: value" appended. Only string values are appended.
-    """
-    updated_msg = msg
-    for k, v in kwargs.iteritems():
-        if isinstance(v, basestring):
-            updated_msg += "\n" + k + ": " + v
-
-    return updated_msg
 
 
 class NiftiMapsMasker(BaseEstimator, TransformerMixin, CacheMixin):
